feat) 추가 연습 문제 마이페이지 저장 기능 구현

+++ src/pages/UserPageEx.jsx
@@ -16,14 +16,18 @@ const UserPageEx = () => {
 
   useEffect(() => {
     // 로컬 스토리지에서 템플릿 데이터 불러오기
-    const storedTemplates = JSON.parse(localStorage.getItem("templates")) || [];
+    const storedTemplates =
+      JSON.parse(localStorage.getItem("practiceTemplates")) || [];
     setTemplates(storedTemplates);
   }, []);
 
-  const handleDeleteTemplate = (templateId) => {
-    const updatedTemplates = templates.filter(
-      (template) => template.templateId !== templateId
+  const handleDeleteTemplate = (id) => {
+    const confirmDelete = window.confirm(
+      "정말로 이 연습문제를 삭제하시겠습니까?"
     );
+    if (!confirmDelete) return;
+
+    const updatedTemplates = templates.filter((template) => template.id !== id);
     setTemplates(updatedTemplates);
     localStorage.setItem("templates", JSON.stringify(updatedTemplates));
     alert("추가 연습 문제 템플릿이 삭제되었습니다.");
@@ -195,7 +199,7 @@ const UserPageEx = () => {
             <DeleteButton
               onClick={(e) => {
                 e.stopPropagation();
-                handleDeleteTemplate(template.templateId);
+                handleDeleteTemplate(template.id);
               }}
             >
               <img src={DeleteIcon} alt="삭제" />
@@ -204,7 +208,7 @@ const UserPageEx = () => {
               <img src={ExIcon} alt="추가 연습 문제" />
             </IconWrapper>
             <CardText>
-              {template.templateName}
+              {template.title}
               <DateText>생성 날짜: {new Date().toLocaleDateString()}</DateText>
             </CardText>
           </Card>
